refactor(client): centralize base64 helpers in a shared module
`decodeBase64` was duplicated in `client/src/response.ts` and
`client-fetch/src/wrap.ts`; the equivalent `encodeBase64` logic was
inlined in `client/src/payload.ts` (UTF-8 → binary-string → btoa) and
also lived in `wrap.ts`. Extract both into `client/src/base64.ts`,
re-export from the package's public surface, and route every call
site through it so the Node/browser fallback logic doesn't drift.

Addresses PR #104 review comment r3319923829.
