+++ typescript/packages/client/src/payload.ts
@@ -3,8 +3,9 @@
 // Defense-in-depth: the assembled `EscrowPaymentPayload` is re-validated
 // through `parseEscrowPaymentPayload` from `@bosonprotocol/x402-core`
 // before serialization, so shape bugs surface here instead of at the
-// server. Base64 encoding picks `Buffer` on Node and `btoa` on browser
-// targets — tsup builds both, so the runtime branch matters.
+// server. Base64 encoding goes through the isomorphic `encodeBase64`
+// helper from `./base64.js` — `Buffer` on Node, UTF-8-aware
+// `btoa` on browsers.
 
 import {
   parseEscrowPaymentPayload,
@@ -16,6 +17,7 @@ import {
 } from "@bosonprotocol/x402-core/schemes/escrow";
 import type { Address } from "viem";
 
+import { encodeBase64 } from "./base64.js";
 import type { ResolvedFulfillment } from "./fulfillment.js";
 
 /** Current x402 protocol version embedded in the payload envelope. */
@@ -79,21 +81,5 @@ export function assemblePayload({
 /** Build, validate, and base64-encode the payload for the `X-PAYMENT` header. */
 export function assembleAndEncodePayload(args: AssembleArgs): string {
   const payload = assemblePayload(args);
-  const json = JSON.stringify(payload);
-  if (typeof Buffer !== "undefined") {
-    return Buffer.from(json, "utf8").toString("base64");
-  }
-  // Browser fallback. Wire payloads mostly carry hex/numeric strings,
-  // but atomic Flow B's `fulfillment.data` is a `Record<string, unknown>`
-  // the buyer populates — emails, addresses, free-form notes. `btoa`
-  // accepts only a binary string (code units 0–255) and throws
-  // `InvalidCharacterError` on any character above U+00FF. UTF-8 encode
-  // the JSON first, then map the bytes into the binary-string form
-  // `btoa` expects.
-  const bytes = new TextEncoder().encode(json);
-  let binary = "";
-  for (const b of bytes) {
-    binary += String.fromCharCode(b);
-  }
-  return btoa(binary);
+  return encodeBase64(JSON.stringify(payload));
 }
